+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - fix "incoming message background color" being used for quotes of outgoing sticker messages #4456
 - fix stickers being smaller than they're supposed to be #4432
 - fix reactions to sticker messages overlapping with next message #4433
+- fix: "Enter" not adding the first contact in "Add Members" dialog #4439
 
 ## [1.50.1] - 2024-12-18